Only run `npm publish` CI workflow on tagged commits
Running the workflow on every push is wasteful, because we know it will
fail (a new version is published to npm only if "version" in
package.json has been changed since last publish). The CI doesn't even
run any tests yet, so it's useless for normal commits.

The commits with the intent to publish a new version should be properly
tagged, so let's run the publish job only on version tags.

If the tag pattern turns out to be too strict in the future, we can
relax it, but the idea is to be conscious about how does a valid version
tag we're using look like (so that a non-version tag isn't automatically
published to npm).