Use named (not anonymous) classes for custom error types
This fixes the `name` property of class constructors. For example,
`KaitaiStream.ValidationNotEqualError.name` now returns
`'ValidationNotEqualError'`. This is consistent with how the built-in
error types work - for example, `RangeError.name` is `'RangeError'`, as
you would expect.

Previously, when the classes were anonymous,
`KaitaiStream.ValidationNotEqualError.name` returned `'class_4'` (which
was an automatically generated name that the TypeScript compiler used in
the output ES5 code). It's worth noting that error constructors were
anonymous also in our handwritten `KaitaiStream.js` before we migrated
to TypeScript ( `KaitaiStream.ValidationNotEqualError.name` used to be
the empty string `''`), so keeping them anonymous wouldn't really be a
regression, but it's nice to improve it now.

I noticed this issue because we use
[`assert.throws`](https://nodejs.org/docs/latest-v20.x/api/assert.html#assertthrowsfn-error-message)
in KST with arguments as in the "Validate instanceof using constructor:"
example. If this assertion fails, Node.js includes the `name` property
of the constructor in the error message if it's not empty. As a result,
the error message was the following when a `ValidationNotEqualError` was
expected but not thrown:

```
Missing expected exception (class_4).
AssertionError [ERR_ASSERTION]: Missing expected exception (class_4).
    at /tests/helpers/javascript/testHelperThrows.js:12:16
    at FSReqCallback.readFileAfterClose [as oncomplete] (node:internal/fs/read/context:68:3)
```

Now it is `Missing expected exception (ValidationNotEqualError)`.