package.json: remove deprecated `prepublish` script
The status of `prepublish` scripts is still unclear and confusing to me.
https://docs.npmjs.com/cli/v10/using-npm/scripts#life-cycle-scripts
states "prepublish (DEPRECATED)", then
https://docs.npmjs.com/cli/v10/using-npm/scripts#prepare-and-prepublish
explains how `prepublish` is a confusing name (because it doesn't
actually run with `publish` at all), but then it still recommends "use a
`prepublish` script" for tasks such as "Compiling CoffeeScript source
code into JavaScript". That doesn't sound like it's deprecated to me if
it's still recommended in some situations, so maybe the documentation is
wrong.

I suppose the least surprising option is not to rely on any automatic
"hook" and always require running `npm run build` manually.