Update typescript and devDependencies
This creates a bit more concise js output from TypeScript.

My main intention was to remove rollup from this project as tsc supports
umd output too. But for some reason the umd output doesnt seem to work
with the tests as the import seems to be different with rollup.
Dropping support for old JS and using ES modules rather than umd would
simplify the toolchain here.
ES modules are supported from Node.js 12 which is also many years out of
support already. All browsers support ES modules for years now so
requiring ES modules shouldnt be a big problem. But for this PR I stayed
with bumping the dependencies.